Piccin + Sommariva
Creata icona di test e comprensione grandezza dell'immagine futura.
Creata ed importata classe Finestra, dove verrà implementata tutta la parte grafica.

Controllo Finale: FUNZIONANTE
