+++ todolist/ToDoList.java
@@ -1,22 +1,26 @@
 package ToDoList;
 import java.io.IOException;
 import java.util.Scanner;
-import mainclasses.Utente;
+
+import Grafica.Finestra;
+import Backhand.Utente;
 
 //Import Grafica
-import javax.swing.JFrame;
+
+
 public class ToDoList {
     public static void main(String[] args) throws IOException {
 
-        //Codice Grafica
-        JFrame frame  = new JFrame();                                           //Crea il frame
-        frame.setTitle("ToDo_List");                                            //Nome frame
-        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);                   //Quando chiudo il frame con la X la chiudo definitivamente (Non la nascondo)
-        frame.setSize(600,600);                                                 //Imposta la dimesione del frame
-        frame.setVisible(true);
+
+        //Finestra                                                                                                      //Creata classe finestra con all'interno tutta la parte grafica
+        Finestra fin = new Finestra();
+        fin.usaFrame();
 
 
 
+
+
+        //Parte funzionale
         Utente utente = new Utente();
         Scanner sc = new Scanner(System.in);
         String cmd = "";
@@ -31,7 +35,7 @@ public class ToDoList {
                     utente.runCmd(cmd);
             }
         }
-        
+
     }
-    
+
 }
